refactor slack-bolt-app code
