+++ packages/slack-bolt-app/src/async-handler.ts
@@ -2,6 +2,8 @@ import { Handler } from 'aws-lambda';
 import { App, AwsLambdaReceiver, LogLevel } from '@slack/bolt';
 import z from 'zod';
 import { getOrCreateWorkerInstance } from '@remote-swe-agents/agent-core/lib';
+import { makeIdempotent } from './util/idempotency';
+import { IdempotencyAlreadyInProgressError, IdempotencyConfig } from '@aws-lambda-powertools/idempotency';
 
 const BotToken = process.env.BOT_TOKEN!;
 
@@ -34,23 +36,33 @@ export const handler: Handler<unknown> = async (rawEvent, context) => {
   const event = eventSchema.parse(rawEvent);
   if (event.type == 'ensureInstance') {
     try {
-      const res = await getOrCreateWorkerInstance(event.workerId, event.slackChannelId, event.slackThreadTs);
+      // When the handler is invoked more than once in short interval,
+      // the second invocation launches another instance because
+      // DescribeInstances does not return the instance launched from
+      // the first invocation very soon. To avoid it, we use makeIdempotent here.
+      await makeIdempotent(
+        async (_: string) => {
+          const res = await getOrCreateWorkerInstance(event.workerId, event.slackChannelId, event.slackThreadTs);
 
-      if (res.oldStatus == 'stopped') {
-        await app.client.chat.postMessage({
-          channel: event.slackChannelId,
-          thread_ts: event.slackThreadTs,
-          text: `Waking up from sleep mode...`,
-        });
-      } else if (res.oldStatus == 'terminated') {
-        await app.client.chat.postMessage({
-          channel: event.slackChannelId,
-          thread_ts: event.slackThreadTs,
-          text: `Preparing for a new instance${res.usedCache ? ' (using a cached AMI)' : ''}...`,
-        });
-      }
+          if (res.oldStatus == 'stopped') {
+            await app.client.chat.postMessage({
+              channel: event.slackChannelId,
+              thread_ts: event.slackThreadTs,
+              text: `Waking up from sleep mode...`,
+            });
+          } else if (res.oldStatus == 'terminated') {
+            await app.client.chat.postMessage({
+              channel: event.slackChannelId,
+              thread_ts: event.slackThreadTs,
+              text: `Preparing for a new instance${res.usedCache ? ' (using a cached AMI)' : ''}...`,
+            });
+          }
+        },
+        { config: new IdempotencyConfig({ expiresAfterSeconds: 30 }) }
+      )(`ensureInstance-${event.workerId}`);
     } catch (e) {
       console.error(e);
+      if (e instanceof IdempotencyAlreadyInProgressError) return;
       await app.client.chat.postMessage({
         channel: event.slackChannelId,
         thread_ts: event.slackThreadTs,
